Move the parameterization functionality into iris_common.h and iris_common.cc.
