+++ planning/iris/BUILD.bazel
@@ -46,6 +46,7 @@ drake_cc_library(
     deps = [
         "//common:name_value",
         "//geometry:meshcat",
+        "//multibody/rational:rational_forward_kinematics",
         "//solvers:mathematical_program",
     ],
 )
