+++ planning/iris/iris_common.h
@@ -9,6 +9,7 @@
 #include "drake/common/name_value.h"
 #include "drake/common/parallelism.h"
 #include "drake/geometry/meshcat.h"
+#include "drake/multibody/rational/rational_forward_kinematics.h"
 #include "drake/solvers/mathematical_program.h"
 
 namespace drake {
@@ -16,6 +17,8 @@ namespace planning {
 
 class CommonSampledIrisOptions {
  public:
+  DRAKE_DEFAULT_COPY_AND_MOVE_AND_ASSIGN(CommonSampledIrisOptions);
+
   /** Passes this object to an Archive.
   Refer to @ref yaml_serialization "YAML Serialization" for background.
   Note: This only serializes options that are YAML built-in types. */
@@ -135,6 +138,100 @@ class CommonSampledIrisOptions {
   @note Internally, these constraints are checked after collisions checking is
   performed. */
   const solvers::MathematicalProgram* prog_with_additional_constraints{};
+};
+
+class IrisParameterizationFunction {
+ public:
+  DRAKE_DEFAULT_COPY_AND_MOVE_AND_ASSIGN(IrisParameterizationFunction);
+
+  typedef std::function<Eigen::VectorXd(const Eigen::VectorXd&)>
+      ParameterizationFunction;
+
+  /** Ordinarily, IRIS algorithms grow collision free regions in the robot's
+   * configuration space C. This allows the user to specify a function f:Q→C ,
+   * and grow the region in Q instead. The function should be a map R^m to
+   * R^n, where n is the dimension of the plant configuration space, determined
+   * via `checker.plant().num_positions()` and m is `parameterization_dimension`
+   * if specified. The user must provide `parameterization`, which is the
+   * function f, `parameterization_is_threadsafe`, which is whether or not
+   * `parametrization` can be called concurrently, and
+   * `parameterization_dimension`, the dimension of the input space Q. */
+  void set_parameterization(const ParameterizationFunction& parameterization,
+                            bool parameterization_is_threadsafe,
+                            int parameterization_dimension) {
+    parameterization_ = parameterization;
+    parameterization_is_threadsafe_ = parameterization_is_threadsafe;
+    parameterization_dimension_ = parameterization_dimension;
+  }
+
+  /** Default constructor -- returns the identity mapping, which is threadsafe
+   * and compatible with any dimension configuration space. */
+  IrisParameterizationFunction() = default;
+
+  /** Alternative to `set_parameterization` that allows the user to define the
+   * parameterization using a `VectorX<Expression>`. The user must also provide
+   * a vector containing the variables used in `expression_parameterization`, in
+   * the order that they should be evaluated. Each `Variable` in `variables`
+   * must be used, each `Variable` used in `expression_parameterization` must
+   * appear in `variables`, and there must be no duplicates in `variables`.
+   * @note Expression parameterizations are always threadsafe.
+   * @throws if the number of variables used across
+   * `expression_parameterization` does not match `ssize(variables)`.
+   * @throws if any variables in `expression_parameterization` are not listed in
+   * `variables`.
+   * @throws if any variables in `variables` are not used anywhere in
+   * `expression_parameterization`. */
+  void SetParameterizationFromExpression(
+      const Eigen::VectorX<symbolic::Expression>& expression_parameterization,
+      const Eigen::VectorX<symbolic::Variable>& variables);
+
+  /** Get the parameterization function.
+   * @note If the user has not specified this with `set_parameterization()`,
+   * then the default value of `parameterization_` is the identity function,
+   * indicating that the regions should be grown in the full configuration space
+   * (in the standard coordinate system). */
+  const ParameterizationFunction& get_parameterization() const {
+    return parameterization_;
+  }
+
+  /** Returns whether or not the user has specified the parameterization to be
+   * threadsafe.
+   * @note The default `parameterization_` is the identity function, which is
+   * threadsafe. */
+  bool get_parameterization_is_threadsafe() const {
+    return parameterization_is_threadsafe_;
+  }
+
+  /** Returns what the user has specified as the input dimension for the
+   * parameterization function, or std::nullopt if it has not been set. A
+   * std::nullopt value indicates that
+   * IrisZo should use the ambient configuration space dimension as the input
+   * dimension to the parameterization. */
+  std::optional<int> get_parameterization_dimension() const {
+    return parameterization_dimension_;
+  }
+
+  /** Constructs an instance of IrisParameterizationFunction that handles a
+   * rational kinematic parameterization. Regions are grown in the `s`
+   * variables, so as to minimize collisions in the `q` variables. See
+   * RationalForwardKinematics for details.
+   * @note The user is responsible for ensuring `kin` (and the underlying
+   * MultibodyPlant it is built on) is kept alive. If that object is deleted,
+   * then the parametrization can no longer be used. */
+  static IrisParameterizationFunction
+  CreateWithRationalKinematicParameterization(
+      const multibody::RationalForwardKinematics* kin,
+      const Eigen::Ref<const Eigen::VectorXd>& q_star_val);
+
+ private:
+  bool parameterization_is_threadsafe_{true};
+
+  std::optional<int> parameterization_dimension_{std::nullopt};
+
+  std::function<Eigen::VectorXd(const Eigen::VectorXd&)> parameterization_{
+      [](const Eigen::VectorXd& q) -> Eigen::VectorXd {
+        return q;
+      }};
 };
 
 }  // namespace planning
