+++ planning/iris/iris_zo.h
@@ -47,88 +47,10 @@ class IrisZoOptions {
   /** Maximum number of bisection steps. */
   int bisection_steps{10};
 
-  typedef std::function<Eigen::VectorXd(const Eigen::VectorXd&)>
-      ParameterizationFunction;
-
-  /** Ordinarily, IRIS-ZO grows collision free regions in the robot's
-   * configuration space C. This allows the user to specify a function f:Q→C ,
-   * and grow the region in Q instead. The function should be a map R^m to
-   * R^n, where n is the dimension of the plant configuration space, determined
-   * via `checker.plant().num_positions()` and m is `parameterization_dimension`
-   * if specified. The user must provide `parameterization`, which is the
-   * function f, `parameterization_is_threadsafe`, which is whether or not
-   * `parametrization` can be called concurrently, and
-   * `parameterization_dimension`, the dimension of the input space Q. */
-  void set_parameterization(const ParameterizationFunction& parameterization,
-                            bool parameterization_is_threadsafe,
-                            int parameterization_dimension) {
-    parameterization_ = parameterization;
-    parameterization_is_threadsafe_ = parameterization_is_threadsafe;
-    parameterization_dimension_ = parameterization_dimension;
-  }
-
-  /** Alternative to `set_parameterization` that allows the user to define the
-   * parameterization using a `VectorX<Expression>`. The user must also provide
-   * a vector containing the variables used in `expression_parameterization`, in
-   * the order that they should be evaluated. Each `Variable` in `variables`
-   * must be used, each `Variable` used in `expression_parameterization` must
-   * appear in `variables`, and there must be no duplicates in `variables`.
-   * @note Expression parameterizations are always threadsafe.
-   * @throws if the number of variables used across
-   * `expression_parameterization` does not match `ssize(variables)`.
-   * @throws if any variables in `expression_parameterization` are not listed in
-   * `variables`.
-   * @throws if any variables in `variables` are not used anywhere in
-   * `expression_parameterization`. */
-  void SetParameterizationFromExpression(
-      const Eigen::VectorX<symbolic::Expression>& expression_parameterization,
-      const Eigen::VectorX<symbolic::Variable>& variables);
-
-  /** Get the parameterization function.
-   * @note If the user has not specified this with `set_parameterization()`,
-   * then the default value of `parameterization_` is the identity function,
-   * indicating that the regions should be grown in the full configuration space
-   * (in the standard coordinate system). */
-  const ParameterizationFunction& get_parameterization() const {
-    return parameterization_;
-  }
-
-  /** Returns whether or not the user has specified the parameterization to be
-   * threadsafe.
-   * @note The default `parameterization_` is the identity function, which is
-   * threadsafe. */
-  bool get_parameterization_is_threadsafe() const {
-    return parameterization_is_threadsafe_;
-  }
-
-  /** Returns what the user has specified as the input dimension for the
-   * parameterization function, or std::nullopt if it has not been set. A
-   * std::nullopt value indicates that
-   * IrisZo should use the ambient configuration space dimension as the input
-   * dimension to the parameterization. */
-  std::optional<int> get_parameterization_dimension() const {
-    return parameterization_dimension_;
-  }
-
-  /** Constructs an instance of IrisZoOptions that handles a rational kinematic
-   * parameterization. Regions are grown in the `s` variables, so as to minimize
-   * collisions in the `q` variables. See RationalForwardKinematics for details.
-   * @note The user is responsible for ensuring `kin` (and the underlying
-   * MultibodyPlant it is built on) is kept alive. If that object is deleted,
-   * then the parametrization can no longer be used. */
-  static IrisZoOptions CreateWithRationalKinematicParameterization(
-      const multibody::RationalForwardKinematics* kin,
-      const Eigen::Ref<const Eigen::VectorXd>& q_star_val);
-
- private:
-  bool parameterization_is_threadsafe_{true};
-
-  std::optional<int> parameterization_dimension_{std::nullopt};
-
-  std::function<Eigen::VectorXd(const Eigen::VectorXd&)> parameterization_{
-      [](const Eigen::VectorXd& q) -> Eigen::VectorXd {
-        return q;
-      }};
+  /** Parameterization of the subspace along which to grow the region. Default
+   * is the identity parameterization, corresponding to growing regions in the
+   * ordinary configuration space. */
+  IrisParameterizationFunction parameterization{};
 };
 
 /** The IRIS-ZO (Iterative Regional Inflation by Semidefinite programming - Zero
