+++ planning/iris/test/iris_zo_test.cc
@@ -16,6 +16,7 @@
 #include "drake/geometry/test_utilities/meshcat_environment.h"
 #include "drake/multibody/inverse_kinematics/inverse_kinematics.h"
 #include "drake/multibody/rational/rational_forward_kinematics.h"
+#include "drake/planning/iris/iris_common.h"
 #include "drake/planning/iris/test/iris_test_utilities.h"
 #include "drake/planning/robot_diagram_builder.h"
 #include "drake/planning/scene_graph_collision_checker.h"
@@ -49,25 +50,28 @@ TEST_F(JointLimits1D, JointLimitsBasic) {
 TEST_F(JointLimits1D, JointLimitsWithParameterization) {
   std::vector<IrisZoOptions> vector_of_options;
   vector_of_options.emplace_back();
-  vector_of_options.back().set_parameterization(
+  vector_of_options.back().parameterization.set_parameterization(
       [](const VectorXd& q) -> VectorXd {
         return q;
       },
       /* parameterization_is_threadsafe */ false,
       /* parameterization_dimension */ 1);
-  EXPECT_EQ(vector_of_options.back().get_parameterization_is_threadsafe(),
+  EXPECT_EQ(vector_of_options.back()
+                .parameterization.get_parameterization_is_threadsafe(),
             false);
 
   // Now set the parameterization with parameterization_is_threadsafe set to
   // true.
   vector_of_options.emplace_back();
-  vector_of_options.back().set_parameterization(
+  vector_of_options.back().parameterization.set_parameterization(
       [](const VectorXd& q) -> VectorXd {
         return q;
       },
       /* parameterization_is_threadsafe */ true,
       /* parameterization_dimension */ 1);
-  EXPECT_EQ(vector_of_options[1].get_parameterization_is_threadsafe(), true);
+  EXPECT_EQ(vector_of_options[1]
+                .parameterization.get_parameterization_is_threadsafe(),
+            true);
 
   // Now use an Expression parameterization.
   Eigen::VectorX<symbolic::Variable> variables(1);
@@ -75,19 +79,23 @@ TEST_F(JointLimits1D, JointLimitsWithParameterization) {
   Eigen::VectorX<symbolic::Expression> parameterization_expression(1);
   parameterization_expression[0] = symbolic::Expression(variables[0]);
   vector_of_options.emplace_back();
-  vector_of_options.back().SetParameterizationFromExpression(
+  vector_of_options.back().parameterization.SetParameterizationFromExpression(
       parameterization_expression, variables);
   // Expression parameterizations are always threadsafe.
-  EXPECT_EQ(vector_of_options.back().get_parameterization_is_threadsafe(),
+  EXPECT_EQ(vector_of_options.back()
+                .parameterization.get_parameterization_is_threadsafe(),
             true);
 
   for (auto& options : vector_of_options) {
     options.sampled_iris_options.verbose = true;
 
     // Check that the parameterization was set correctly.
-    ASSERT_TRUE(options.get_parameterization_dimension().has_value());
-    EXPECT_EQ(options.get_parameterization_dimension().value(), 1);
-    const Vector1d output = options.get_parameterization()(Vector1d(3.0));
+    ASSERT_TRUE(
+        options.parameterization.get_parameterization_dimension().has_value());
+    EXPECT_EQ(options.parameterization.get_parameterization_dimension().value(),
+              1);
+    const Vector1d output =
+        options.parameterization.get_parameterization()(Vector1d(3.0));
     EXPECT_NEAR(output[0], 3.0, 1e-15);
 
     HPolyhedron region =
@@ -112,7 +120,7 @@ TEST_F(JointLimits1D, ParameterizationExpressionErrorChecks) {
       symbolic::Expression(variables[0]);
   parameterization_expression_wrong_dimension[1] =
       symbolic::Expression(variables[0]);
-  options.SetParameterizationFromExpression(
+  options.parameterization.SetParameterizationFromExpression(
       parameterization_expression_wrong_dimension, variables);
   DRAKE_EXPECT_THROWS_MESSAGE(
       IrisZo(*checker_, starting_ellipsoid_, domain_, options),
@@ -124,13 +132,13 @@ TEST_F(JointLimits1D, ParameterizationExpressionErrorChecks) {
   Eigen::VectorX<symbolic::Expression>
       parameterization_expression_extra_variable(1);
   parameterization_expression_extra_variable[0] = variables[0] + extra_variable;
-  EXPECT_THROW(options.SetParameterizationFromExpression(
+  EXPECT_THROW(options.parameterization.SetParameterizationFromExpression(
                    parameterization_expression_extra_variable, variables),
                std::exception);
   Eigen::VectorX<symbolic::Expression>
       parameterization_expression_missing_variable(1);
   parameterization_expression_missing_variable[0] = symbolic::Expression(1);
-  EXPECT_THROW(options.SetParameterizationFromExpression(
+  EXPECT_THROW(options.parameterization.SetParameterizationFromExpression(
                    parameterization_expression_missing_variable, variables),
                std::exception);
 }
@@ -153,8 +161,9 @@ TEST_F(DoublePendulum, IrisZoTest) {
 // configuration space built from RationalForwardKinematics.
 TEST_F(DoublePendulumRationalForwardKinematics,
        ParameterizationFromStaticConstructor) {
-  IrisZoOptions options =
-      IrisZoOptions::CreateWithRationalKinematicParameterization(
+  IrisZoOptions options;
+  options.parameterization =
+      IrisParameterizationFunction::CreateWithRationalKinematicParameterization(
           &rational_kinematics_,
           /* q_star_val */ Vector2d::Zero());
   options.sampled_iris_options.verbose = true;
@@ -163,11 +172,14 @@ TEST_F(DoublePendulumRationalForwardKinematics,
   options.sampled_iris_options.meshcat = meshcat_;
 
   // Check that the parameterization was set correctly.
-  EXPECT_EQ(options.get_parameterization_is_threadsafe(), true);
-  ASSERT_TRUE(options.get_parameterization_dimension().has_value());
-  EXPECT_EQ(options.get_parameterization_dimension().value(), 2);
+  EXPECT_EQ(options.parameterization.get_parameterization_is_threadsafe(),
+            true);
+  ASSERT_TRUE(
+      options.parameterization.get_parameterization_dimension().has_value());
+  EXPECT_EQ(options.parameterization.get_parameterization_dimension().value(),
+            2);
 
-  CheckParameterization(options.get_parameterization());
+  CheckParameterization(options.parameterization.get_parameterization());
 
   HPolyhedron region =
       IrisZo(*checker_, starting_ellipsoid_rational_forward_kinematics_,
@@ -175,7 +187,8 @@ TEST_F(DoublePendulumRationalForwardKinematics,
 
   CheckRegionRationalForwardKinematics(region);
   PlotEnvironmentAndRegionRationalForwardKinematics(
-      region, options.get_parameterization(), region_query_point_1_);
+      region, options.parameterization.get_parameterization(),
+      region_query_point_1_);
 }
 
 // Verify that we can get the same behavior by using an Expression
@@ -193,12 +206,12 @@ TEST_F(DoublePendulumRationalForwardKinematics,
     parameterization_expression[i] =
         atan2(2 * variables[i], 1 - pow(variables[i], 2));
   }
-  options.SetParameterizationFromExpression(parameterization_expression,
-                                            variables);
-  EXPECT_TRUE(options.get_parameterization_is_threadsafe());
-  EXPECT_EQ(options.get_parameterization_dimension(), 2);
+  options.parameterization.SetParameterizationFromExpression(
+      parameterization_expression, variables);
+  EXPECT_TRUE(options.parameterization.get_parameterization_is_threadsafe());
+  EXPECT_EQ(options.parameterization.get_parameterization_dimension(), 2);
 
-  CheckParameterization(options.get_parameterization());
+  CheckParameterization(options.parameterization.get_parameterization());
 
   HPolyhedron region =
       IrisZo(*checker_, starting_ellipsoid_rational_forward_kinematics_,
@@ -210,7 +223,7 @@ TEST_F(DoublePendulumRationalForwardKinematics, BadParameterization) {
   // Verify that we fail gracefully if the parameterization has the wrong output
   // dimension (even if we claim it outputs the correct dimension).
   IrisZoOptions options;
-  options.set_parameterization(
+  options.parameterization.set_parameterization(
       [](const VectorXd& q) -> VectorXd {
         return Vector1d(0.0);
       },
@@ -319,7 +332,7 @@ TEST_F(ConvexConfigurationSpaceWithThreadsafeConstraint, BadContainmentPoint) {
 TEST_F(ConvexConfigurationSubspace, AdditionalConstraintsDimensionMismatch) {
   IrisZoOptions options;
 
-  options.set_parameterization(
+  options.parameterization.set_parameterization(
       [](const Vector1d& config) -> Vector2d {
         return Vector2d{config[0], 2 * config[0] + 1};
       },
@@ -339,7 +352,7 @@ TEST_F(ConvexConfigurationSubspace, AdditionalConstraintsDimensionMismatch) {
 
 TEST_F(ConvexConfigurationSubspace, FunctionParameterization) {
   IrisZoOptions options;
-  options.set_parameterization(
+  options.parameterization.set_parameterization(
       [](const Vector1d& config) -> Vector2d {
         return Vector2d{config[0], 2 * config[0] + 1};
       },
@@ -350,7 +363,8 @@ TEST_F(ConvexConfigurationSubspace, FunctionParameterization) {
   CheckRegion(region);
 
   meshcat_->Delete();
-  PlotEnvironmentAndRegionSubspace(region, options.get_parameterization());
+  PlotEnvironmentAndRegionSubspace(
+      region, options.parameterization.get_parameterization());
 }
 
 TEST_F(ConvexConfigurationSubspace, ExpressionParameterization) {
@@ -365,10 +379,10 @@ TEST_F(ConvexConfigurationSubspace, ExpressionParameterization) {
   parameterization_expression[1] = 2 * symbolic::Expression(variables[0]) + 1;
 
   IrisZoOptions options;
-  options.SetParameterizationFromExpression(parameterization_expression,
-                                            variables);
-  EXPECT_TRUE(options.get_parameterization_is_threadsafe());
-  EXPECT_EQ(options.get_parameterization_dimension(), 1);
+  options.parameterization.SetParameterizationFromExpression(
+      parameterization_expression, variables);
+  EXPECT_TRUE(options.parameterization.get_parameterization_is_threadsafe());
+  EXPECT_EQ(options.parameterization.get_parameterization_dimension(), 1);
 
   HPolyhedron region = IrisZo(*checker_, starting_ellipsoid_, domain_, options);
   CheckRegion(region);
